implement feedback from PR review
