+++ .gitignore
@@ -34,11 +34,7 @@ captures/
 
 # Intellij
 *.iml
-.idea/workspace.xml
-.idea/tasks.xml
-.idea/gradle.xml
-.idea/dictionaries
-.idea/libraries
+.idea/
 
 # Keystore files
 *.jks
@@ -53,3 +49,5 @@ google-services.json
 freeline.py
 freeline/
 freeline_project_description.json
+libs/
+app/libs/
