--- .idea/assetWizardSettings.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="WizardSettings">
-    <option name="children">
-      <map>
-        <entry key="imageWizard">
-          <value>
-            <PersistentState />
-          </value>
-        </entry>
-      </map>
-    </option>
-  </component>
-</project>
